renamed modules in prior_model package
