+++ autofit/mapper/__init__.py
@@ -1,7 +1,7 @@
 from autofit.mapper.prior_model.prior_model import PriorModel
 from autofit.mapper.prior_model.annotation import AnnotationPriorModel
-from autofit.mapper.prior_model.collection_prior_model import CollectionPriorModel
-from autofit.mapper.prior_model.abstract_prior_model import AbstractPriorModel
+from autofit.mapper.prior_model.collection import CollectionPriorModel
+from autofit.mapper.prior_model.abstract import AbstractPriorModel
 from .model_mapper import ModelMapper
 from .prior import Prior, UniformPrior, GaussianPrior, Constant, LogUniformPrior, \
     AttributeNameValue
