+++ autofit/mapper/prior_model/abstract.py
