+++ test/integration/mapper.py
@@ -1,14 +1,14 @@
 import numpy as np
 import pytest
 
-import autofit.mapper.prior_model.collection_prior_model
+import autofit.mapper.prior_model.collection
 from autofit import mapper
 from test import mock
 
 
 @pytest.fixture(name="source_light_profiles")
 def make_source_light_profiles():
-    source_light_profiles = autofit.mapper.prior_model.collection_prior_model.CollectionPriorModel(
+    source_light_profiles = autofit.mapper.prior_model.collection.CollectionPriorModel(
         light=mock.EllipticalLP
     )
     return source_light_profiles
@@ -26,10 +26,10 @@ def make_source(source_light_profiles):
 def make_tracer_prior_model(source):
     lens = mapper.PriorModel(
         mock.Galaxy,
-        light_profiles=autofit.mapper.prior_model.collection_prior_model.CollectionPriorModel(
+        light_profiles=autofit.mapper.prior_model.collection.CollectionPriorModel(
             light=mock.EllipticalLP
         ),
-        mass_profiles=autofit.mapper.prior_model.collection_prior_model.CollectionPriorModel(
+        mass_profiles=autofit.mapper.prior_model.collection.CollectionPriorModel(
             mass=mock.EllipticalMassProfile
         )
     )
